backend/requirements.txt
fastapi==0.135.1
uvicorn[standard]==0.41.0
sqlalchemy==2.0.48
psycopg2-binary==2.9.11
pydantic-settings==2.13.1
beautifulsoup4==4.14.3
httpx==0.28.1
